Update patch set 7
Patch Set 7:

There is a discussion on DPTF v.s. External sensor. If we use an external sensor, we may need  to further tune the fans. Waiting for the discussion to be closed. Other than that, it's LGTM.

Patch-set: 7
Reviewer: Gerrit User 1106949 <1106949@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>