Update patch set 12
Patch Set 12:

> Patch Set 12:
> 
> > Patch Set 12:
> > 
> > This patch takes the file ec.bin (by default) as input data to program SPI flash.
> > The older one takes ec.RO.flat and ec.RW.bin and program them separately.
> > The size of these images built for cheza on TOT:
> > ec.bin     - 624928 bytes
> > ec.RO.flat - 101080 
> > ec.RW.bin  - 100640 
> > 
> > It requires about three times the time to program the whole ec.bin than to program RO+RW separately.
> > 
> > However, you can still program the RO and RW separately by:
> > > ./util/flash_ec --board=cheza --ro
> > > ./util/flash_ec --board=cheza --image build/cheza/RW/ec.RW.bin  --offset 0x80000
> > 
> > The following is the time spent to program flash before/after this patch for your reference:
> >  - After this patch 
> >    ec.bin : (./util/flash_ec --board=cheza)
> >      real    1m24.916s
> >      user    0m2.132s
> >      sys     0m7.888s
> >    ec.RO.flat : (./util/flash_ec --board=cheza --ro)
> >      real    0m15.427s
> >      user    0m1.244s
> >      sys     0m1.516s
> >    ec.RW.bin : (./util/flash_ec --board=cheza --image build/cheza/RW/ec.RW.bin  --offset 
> >                 0x80000)
> >      real    0m15.232s
> >      user    0m1.208s
> >      sys     0m1.496s
> > - Before this patch
> >      real    0m28.479s
> >      user    0m1.348s
> >      sys     0m2.632s
> 
> ha it's a terrible regression ...
> it seems the new version should just skip writing empty flash chunks since this seems to be the root cause.

Hi Vincent:
The main purpose of this CL is to solve inconvenience when ODM wants to upload ec firmware but only receives the single image file (ec.bin). Or they need to split it into ec.RO.flat and ec.RW.bin manually.

Regarding skipping empty flash chunks, it might leave the previous firmware data in ec's internal flash. For example, if previous firmware image size is 5KB and the current one is 3KB, we will leave the other 1KB at address 0x1000. (The sector of internal flash in npcx ec is 4KB.) Is it OK to you? If so, we will push a new version to skip empty chunks for better performance.

Ps. We can read the flash data through UART to compare the difference. But it might also influence the performance since UART's speed is the bottleneck of uut tool.

Patch-set: 12
Reviewer: Gerrit User 1121210 <1121210@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>