Update patch set 12
Patch Set 12:

> Regarding skipping empty flash chunks, it might leave the previous firmware data in ec's 
> internal flash. For example, if previous firmware image size is 5KB and the current one
>  is 3KB, we will leave the other 1KB at address 0x1000.
> (The sector of internal flash in npcx ec is 4KB.) 
> Is it OK to you? 

No, it is not. It would be even more broken than the current behavior.
I meant skipping the 'write' operation but NOT the 'erase' operation.
Usually uploading the data over UART takes a fair chunk of time.

> Ps. We can read the flash data through UART to compare the difference.
>  But it might also influence the performance since UART's speed is the bottleneck
> of uut tool.

Yes as correctly observed, this will likely be even worse for performance.

Patch-set: 12