cr50: prepare for supporting both x25519 and p256 test keys
The signer script is checking the elf files for presence of test RMA
keys, currently hardcoded to be x25519 keys.

The algorithm (x25519 vs p256) is going to become a compile time
option, the script should be prepared to determine the type of the key
at run time, because the script could be used for signing images from
different branches, compiled with different config options.

The prod p256 key does not yet exist.

BRANCH=none
BUG=b:73296606
TEST=verified that prod signing images including x25519 keys is still
     working as expected.

Change-Id: Icf48845279912ecc9ccdecec1764fcb5f85d22bd
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>