Update patch set 2
Patch Set 2:

(2 comments)

Looking at the proto2 yorp schematic I believe the configuration is set so that:

* In the normal case the hardware advertises CDP (i.e. 7.5W) on both ports
* The first port to claim that power 'wins', its load-switch asserts STATUS_L
* The STATUS_L are crossed between ports, so when one 'wins' the other switches to low current limit and advertising SDP
* Each port also has a software override CHARGE_EN_L, if set high then that port is forced to low current and SDP

In general the software override is not needed (i.e. charging should always be enabled on both ports) and the hardware allocates 2.4A from the 5V rail between the two ports. In some particular system it is possible the software could decide it needs 3W from the 5V rail for some other use (we don't do this but an example would be to allocate to a USB-C port) and thus would need to disable CHARGE_EN_L on both ports. (In the future we could make this even more of a hairball and have only one EC pin to control both CHARGE_EN_L pins, introducing yet another #define...)

Patch-set: 2
CC: Gerrit User 1000094 <1000094@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>