Update patch set 3
Patch Set 3: Code-Review-1

This is going to be in a big conflict with this CL that's underway https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/1072637.

Are you certain that the init bit in battery status has the exact same meaning as the discharge/charge FET status bits? Does this remove support for non-TI fuel gauges?

I think typing the discharge FET status to BP_YES/BP_NO is the wrong direction to go. The FET status check is only relevant to prevent_power_on which is why I had just separated that. This CL would undo that change. As a result, if the battery is in cutoff, it can get stuck in the cutoff state because charge_state_v2 will be in ST_IDLE. Unless of course the FET status isn't relevant, so that problem won't show up, but we haven't protected against the brownout issue that we had on Eve.

Patch-set: 3
Reviewer: Gerrit User 1123814 <1123814@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=-1