Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> Hi Scott: would you like me to abandon this patch as you are planning to merge 
> https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/1072637

I do think the CL I listed needs to land, and if changes are necessary, should be modified on top of that. Given that the fuel gauge access code is moving to /common, we will want to keep that as flexible as possible so that it can more easily support all battery types used.

Patch-set: 4
Reviewer: Gerrit User 1123814 <1123814@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>