Update patch set 4
Patch Set 4:

> Patch Set 3: Code-Review+1
> 
> I think this is ok.

Maybe using the watchdog_reload() for this case is ok 
but IMO the explanation in the CL is not...
It's not triggering the watchdog because the init is taking 1.6s as it is sleeping most of the in the big fat msleep(100) ... it is because this code is called from a deferred function, hence blocking the other HOOKS (including the watchdog kicking).
Therefore it would be nice to document this and also think twice whether we want to block all other HOOKs for 2 seconds (rather than doing something slightly more subtle, e.g. re-calling a deferred every 100ms to do the check in the loop)

Patch-set: 4