Fizz: Move BJ adapter spec table to EC
Currently, the BIOS carries the table which maps (OEM,SKU) to barrel
jack adapter spec. This patch moves this table to the EC. Then, the
EC will independently manage the max voltage and current for BJ.

This would remove the dependency on AP-EC communication, thus improves
the stability

This patch also corrects the mapping between SKUs and BJ wattages.
			SKU	BJ(W)
* KBL-R i7 8550U	4	90
* KBL-R i5 8250U	5	90
* KBL-R i3 8130U	6	90
* KBL-U i7 7600		3	65
* KBL-U i5 7500		2	65
* KBL-U i3 7100		1	65
* KBL-U Celeron 3965	7	65
* KBL-U Celeron 3865	0	65

Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>

BUG=b:109762580
CQ-DEPEND=CL:1089370
BRANCH=none
TEST=Verify BJ adapter is set expectedly on Teemo.

Change-Id: I70c8987670e7495a32fdcbc572779fdc9362e22f