keyboard_scan: Add "ksstate force" to always scan KB matrix
For early bringup, or failure analysis, it is sometimes useful to
be able to force enable the keyboard matrix scanning, even though
other signals (lid close, usb off) would normally disable it.

The only way to disable the scanning again is to wait for an
event that would disable scanning again, or reboot the board, but
that's ok for this debugging purpose.

BRANCH=none
BUG=b:109743721
TEST=Provide power to whiskers via servo only.
     ksstate force => key presses are shown

Change-Id: I3eaa9552ea52f7e3df45fdb6c8d0aa88c7b164b3
Signed-off-by: Nicolas Boichat <drinkcat@chromium.org>