cheza: Mux the 1st-plugged UFP port without USB-capable-bit unset to AP
There are 2 USB-C ports on Cheza. A user may plug a USB-C to USB-A
male cable to one of the ports; then EC should mux the port from USB
hub to AP's primary USB controller, which is configured to do USB
peripheral mode only.

The basic policy is to mux the first-plugged UFP port to AP. However,
a charger port is also a UFP port. In the common scenario that one
port has a charger plugged, plugging a USB-C to A cable fails to
switch the mux (the charger is the first UFP).

If the charger supports PD, we can exclude the charger if its
"USB Communications Capable" bit unset in the Fixed Supply PDO.
But if the charger doesn't support PD, we can't differentiate it
from a USB-C to A cable and the above failure happens. Hopefully,
the official charger does support PD.

BRANCH=none
BUG=b:74395451, b:110265762
TEST=Check the following scenario:
* Plug charger w/ PD to P0, plug C-to-A to P1, check P1 mux to AP.
* Plug charger through hub to P0, plug C-to-A to P1, check P1 mux to AP.
* Continue the above case, boot into kernel and check USB peripheral mode.
* Plug charger w/o PD to P0, plug C-to-A to P1, check P0 mux to AP.
* Swap P0 and P1 for the above cases and check the results.

Change-Id: I4034fa66c0b27cc48d0959bb3f1750690ad5e3f7
Signed-off-by: Wai-Hong Tam <waihong@google.com>