USB PD: Initialize CC polarity during init.
Previously, explicit PD contracts were not maintained after a reset, but
that has recently changed.  Now, if a PD contract is in place, a
SoftReset is sent to the port partner to renogotiate the contract
without dropping VBUS.  However, the CC polarity was not initialized in
this code path.  On a system with external TCPCs, we may have been lucky
and depending upon the orientation, the right CC line may have been
used.  This was actually breaking boards that did lose their TCPC state
after reset.

This commit simply initializes the CC polarity before potentially
sending any PD messages.

BUG=b:111114159
BRANCH=whichever take the new SoftReset after reset patches.
TEST=Flash servo_v4; Plug in PD source to CHG port, reboot, verify that
PD communication still works with the source following a reboot.
TEST=Flash scarlet; Plug in a charge through hub with a charger and USB
storage devices plugged in, reboot, verify that PD communication still
works with the source following a reboot.  Repeat the test in the other
orientation.

Change-Id: I85a16dd8982747a66883579bb8cf3673dbdd95d8
Signed-off-by: Aseda Aboagye <aaboagye@google.com>