util: UUT: skip writing empty segment to save the flash time
In order to reduce the time of programming EC image, the UUT in this CL
checks each segment before it is written. If the content of the segment
is empty (ie. all 0xFF), the UUT will not write the empty data to flash
but only erase it.

This CL also adds a new filed mon_oper_mode in the monitor header.
It is used to indicate which operation the monitor lfw should execute
whenever it is called.

BRANCH=none
BUG=none
TEST=No build errors for make buildall.
TEST=
------------------------------------------------------------------------
1. Connect the servo connector (J24) on npcx7 EVB to servo board v2 via
flex cable.
2. Manually turn the switch SW1.6 to "ON" on npcx7 EVB.
3. Reset ec by issuing Power-Up or VCC1_RST reset.
4. Manually turn the switch SW1.6 to "OFF" on npcx7 EVB.
5. Move npcx7_evb from array BOARDS_NPCX_7M7X_JTAG to BOARDS_NPCX_UUT in
flash_ec.
6. "./util/flash_ec --board=npcx7_evb", make sure ec boots up.
    (Note: "ec_reset" must be removed in line 1051 of flash_ec)
5."time ./util/flash_ec --board=cheza"
     flash time before this CL: 1m26.489s
     flash time after this CL : 0m36.760s
6. Dump the flash content via JTAG, make sure it is the same as cheza's
ec.bin.

Change-Id: Id5ee2bd3a03d13e9b693267b03613c9f2847e0c8
Signed-off-by: CHLin <CHLIN56@nuvoton.com>