npcx: Enable PWM when starting to spin fans
Currently, the EC disables PWM for the fan with PWM=0 but it doesn't
enable PWM when PWM goes up.

This patch ensures the corresponding PWM channel is enabled when the EC
starts spinnning a fan.

This patch also makes the EC disable thermal fan control in S3/S5.

Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>

BUG=b:109951096
BRANCH=none
TEST=Verify the followings on Teemo:
1. Fan stops when RPM=0
2. Fan starts spinning when RPM goes up from 0
3. Fan stops when entering S3/S5
4. Fan starts spinning when leaving S3.

Change-Id: I49654df81953dc22eb1e6f34cb6534886fee6343