host_event: Handle SCI/SMI masks correctly when using S0ix
When host is using S0ix, BIOS is not involved during suspend or resume
paths. Thus, until now, SCI and SMI masks were preserved across S0ix
suspend/resume cycles. However, it was identified in b:78497503 that
if SCI is generated after host enters S0ix, then it results in EC eSPI
controller waiting for host controller to respond to the SCI virtual
wire event. This prevents the EC from entering low power idle mode.

In order to fix the above issue, this change ensures that SCI and SMI
masks are cleared when host expresses its interest to enter
S0ix. Since these masks are not re-programmed by the host on resume,
EC maintains backup copies of both the masks. On resume from S0ix,
backup copies are used to restore the masks for SCI and SMI.

This change applies to both eSPI and LPC systems since SCI/SMI should
not be generated (physical or virtual) when host is in S0ix.

This implementation follows the design proposed in:
https://docs.google.com/document/d/1J2VYeVXV-KZnIHvtDyOrcH4AXJipsSX6m8oZIDSdtv8/edit?usp=sharing

BUG=b:78497503
BRANCH=None
TEST=Verified following:
1. SCI and SMI masks are programmed correctly on boot-up.
2. SCI and SMI masks are set to 0 when host enters S0ix.
3. SCI and SMI masks are restored correctly on S0ix resume.

Change-Id: Iaa3981e6545d2d3cff51649642f6926f0d4ec31f
Signed-off-by: Furquan Shaikh <furquan@google.com>