Update patch set 2
Patch Set 2:

> +jwerner
> 
> We need to remove bootimage USE flag from kevin and gru overlays (so that FW (and EC) does not get build on ToT for these board). Hopefully we are ok with losing FW coverage as well? (I guess scarlet can still provide RK3399 coverage?)

I'm okay if this is a one-off case and we still have Scarlet for 3399. But in general we like to have our ToT coverage for AP firmware, and EC problems shouldn't be a reason to drop it. How did we solve this for other boards... Jerry, Minnie and Elm all still have bootimage set in ToT, do their EC builds also all still work? If this becomes a regular problem we should change the ebuild infrastructure to allow disabling the EC separately somehow.

Patch-set: 2