ppc: sn5s330: Set VCONN ILIM to 0.6A (min).
It seems that several charge through hubs have inrush currents that
exceed 350mA when VCONN is applied.  This in turn causes the SN5S330 to
latch off VCONN due to the overcurrent condition.  This commit simply
sets the VCONN current limit to the 600mA (min) setting.

BUG=b:110370179
BRANCH=None
TEST=Flash nocturne; Use a variety of charge through hubs, verify that
VCONN is not latched off.

Change-Id: I7a0e5d1aa7c42d22b43a596520790debee47163b
Signed-off-by: Aseda Aboagye <aaboagye@google.com>