Nami: Set full charge level to 94%
This patchs sets CONFIG_BATTERY_LEVEL_NEAR_FULL=94.

There are set threshold (100%) and clear threshold (95%). With this
change, the battery LED will show 'full' while the system
keeps charging until the SOC drops below 95%.

So, between 95% and 100%, the battery could be charging or
discharging depending on whether 100% is reached or not.

Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>

BUG=b:109954565
BRANCH=none
TEST=make BOARD=nami

Change-Id: I2009e1d98e6e58e5fac72e363513996909aca764