Nami: Set full charge level to 95%
battery_charge readings are in sync but when the soc gets close to full,
the readings on OS and EC differ. More specifically, OS reading jumps
from 95% to 100%. So, OS reaches 100% but EC is only 95%.

By setting CONFIG_BATTERY_LEVEL_NEAR_FULL=95, we can make the LED in
sync.

Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>

BUG=b:109954565
BRANCH=none
TEST=Verify LED synchronously changes to 'full' with power_supply_info
on Sona.

Change-Id: I2009e1d98e6e58e5fac72e363513996909aca764