Charger: Use display battery percentage to control LED
battery_charge readings get out of sync as the soc gets closer to full
or empty. This causes the LED to show wrong battery status.

This patch converts the actual battery percentage to the display percentage
using the same conversion used by the OS and ACPI to determine whether the
LED should show battery is full or not.

Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>

BUG=b:109954565
BRANCH=none
TEST=Verify LED synchronously changes to 'full' and 'low' with
power_supply_info on Sona.

Change-Id: I2009e1d98e6e58e5fac72e363513996909aca764