Cr50: Add valid PCR value for pinweaver
In order to bind the PIN authentication to PCR4, required for
additional security, a set of valid PCR criteria is added as metadata
in the leaf of the tree. Each criteria has a bitmask of PCR indexes
and the digest sha256 that should be obtained from concatenation of
PCR values for the respective indexes. Pinweaver will validate
the le_secret and release the he_secret only after reading the PCR
values and checking that at least on criteria is satisfied.

For migration of old leaves that don't have the new field, the new
function pw_seal_to_pcr is added. Chrome OS detects when
a leaf is not migrated and calls it instead of pw_handle_try_auth.

For security the migration is not allowed without valid credentials.
So, pw_seal_to_pcr does both, authentication and migration of the
leaf. Also it checks that the leaf was not migrated already.

The PCR criteria set is created on Chrome OS side. Details of that
implementation is in
https://chromium-review.googlesource.com/c/chromiumos/platform2/+/1124856

BRANCH=none
BUG=chromium:812165
TEST=manual

Change-Id: If778c4e46b9945afadfd2af7d58353005624d668
Signed-off-by: igorcov@chromium.org