Cr50: Add valid PCR value for pinweaver
In order to bind the PIN authentication to PCR4, required for
additional security, a valid PCR value is added as metadata in the leaf
of the tree. The pinweaver will validate the le_secret and release the
he_secret only after reading the value of PCR4 and checking that it's
either the default value (unchanged), or the value stored in the leaf.

Each user will have it's own valid PCR4 value and it will be provided
from the caller when insert_leaf is requested.

BRANCH=none
BUG=chromium:812165
TEST=manual

Change-Id: If778c4e46b9945afadfd2af7d58353005624d668
Signed-off-by: igorcov@chromium.org