nocturne: Support new base USB fault protection IC.
There is a new protection IC being introduced that will indicate a fault
on the data lines.  This commit adds support for that new fault pin
which will follow the same behaviour as the other base power fault
pin. The EC should disable pogo power when this goes off.

BUG=b:110204244
BRANCH=None
TEST=make -j BOARD=nocturne

Change-Id: Ia29bbe7109abf22a372e4d50870d870b8bbde4c7
Signed-off-by: Aseda Aboagye <aaboagye@google.com>