test/pinweaver: Fix clang compilation error
When building with clang -Waddress-of-packed-member, it throws a
warning:
test/pinweaver.c:2032:5: error: taking address of packed member 'data_length'
    of class or structure 'pw_request_header_t' may result in an
      unaligned pointer value [-Werror,-Waddress-of-packed-member]
                        &buf.request.header.data_length);

This is because struct pw_test_data_t is defined as __packed (actually,
its members are), which also implies a base alignment of 1 byte.

Tell the compiler that the structure must be aligned on a 4-byte
boundary, which fixes the issue above.

BRANCH=none
BUG=chromium:854924
TEST=make buildall -j
TEST=make CC=clang CFLAGS_y=-Waddress-of-packed-member run-pinweaver -j

Change-Id: I498ff311438303b3f648e370af580075dab613a9
Signed-off-by: Nicolas Boichat <drinkcat@chromium.org>