Update patch set 1
Patch Set 1:

(4 comments)

I haven't run this yet on my system with the wires attached so I can get Saleae traces, but I will do that. I am struggling a bit to see what problems this is solving relative to the solution that's already proposed. I do think the debounce timer of 10 msec makes a self generating enter/exit cycle less likely, but that timer can easily be added to the fix I have currently proposed. Not maintaining a 'lpm_state' was considered a potential weakness, but with the state not tightly coupled to the actual exit, other potential issues are introduced with PD_LPM_FLAGS_ENGAGED gating whether to enter low power mode or not.

Patch-set: 1