careena, grunt: Enable pullup for the lid switch.
In the case that the lid switch becomes loose/disconnected, default to
the 'lid open' state by pulling up the GPIO. Grunt is not affected, but
we want to demonstrate best practices for the grunt family with its
configuration.

BUG=b:110836185
BRANCH=none
TEST=disconnected lid switch on careena and observed lid state remained
open and stable.  Reconnected lid switch and observed lid state followed
physical lid position.

Change-Id: I951594ed7f6d98cf1b5258a894d0b95b6ca98ecc
Signed-off-by: Jonathan Brandmeyer <jbrandmeyer@chromium.org>