nocturne: Turn off sensor interrupts in S5.
To save power, in the next board revisions, the sensors will be disabled
in S5.  This would cause the interrupt lines to be low which would be a
problem since they are active low.

This commit disables the sensor interrupts when entering S5 and then
enables them when entering S0.

BUG=b:109934082
BRANCH=None
TEST=Flash nocturne; power on AP, verify that sensors still work.  Power
cycle AP, verify that sensors still work.

Change-Id: Id1700dca280b06f4cff26c02c308311e0809d559
Signed-off-by: Aseda Aboagye <aaboagye@google.com>