careena: add charger profile override
Currently the charger state is keeping charging state even battery
temperture was ouside of range.
This patch adds charger profile to put charger state to idle if
battery temperture was outside of the allowable the range.

BUG=none
BRANCH=none
TEST=Heat gun on battery then makes it outside of the allowable
temperature range. Verify the charge state is idle with console
command "chgstate".

Change-Id: Ie1eb469548faf1d4849dda94fdf4039fb25d5d6a
Signed-off-by: Devin Lu <Devin.Lu@quantatw.com>