USB-PD: Advertise high current Rp after Vconn is applied
When a device such as HDMI dongle which has Ra/Rd is connected to a CCD
port, CC1 and CC2 voltages fall in the CCD range if we advertise 3A
(Rp=4.7kohm).

This patch makes TCPM first advertise default USB current 900mA (Rp=56kohm)
then advertise 3A (or whatever the higher current the board defines)
after a sink device is attached and Vconn is turned on. Once Vconn is
turned on, CCx voltage will be out of the CCD range.

Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>

BUG=b:110283597
BRANCH=none
TEST=Verify picture is displayed on Akali + HP Z27n when
1. HDMI cable and Dingdong are connected first then DUT is connected
2. Dingdong and DUT are connected first then HDMI cable is connected
Repeat 1 and 2 with the other port connected to a charger.

Change-Id: I2131268abf7ea6414b8277f924e4cd48f6a98800