USB-PD: Advertise high current Rp after Vconn is applied
When a device such as HDMI dongle which has Ra/Rd is connected to a CCD
port, CC1 and CC2 voltages fall in the CCD range if we advertise Rp3A
(=4.7kohm).

This patch makes TCPM first advertise default USB Rp (=56kohm), then
advertise Rp3A (or whatever the higher current the board defines)
after a sink device is attached and Vconn is turned on. Once Vconn is
turned on, CCx voltage will be out of the CCD range.

Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>

BUG=b:110283597
BRANCH=none
TEST=Verify picture is displayed on Vayne + HP Z27n when
1. HDMI cable and Dingdong are connected first then DUT is connected
2. Dingdong and DUT are connected first then HDMI cable is connected

Change-Id: I2131268abf7ea6414b8277f924e4cd48f6a98800