tcpm: it83xx: IT8320 family compatibility
The bit7 and bit3 at UFPVDR register are reserved on BX version.
But in DX version, these two bits are used to get the result of
fast swap voltage comparing in UFP mode. So we change to mask
three bits only.
Add the support of two TX SOP type on the DX version.
(Debug SOP' and Debug SOP'')

BUG=none
BRANCH=none
TEST=Plug USB-C power adapter and USB-C to hdmi adapter, both work.

Change-Id: If881ef54145f211f7d48a971f56a6118487d9eed
Signed-off-by: Dino Li <Dino.Li@ite.com.tw>